move common testing functions into conftest
