+++ .github/workflows/tests.yml
@@ -7,6 +7,7 @@ on:
   pull_request:
     branches: ['*']
 
+# cancel previous runs if new one is triggered 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
   cancel-in-progress: true
